% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{absts_ca1009}
\alias{absts_ca1009}
\title{vector of character strings representing study.abstracts from the 1009 cancer studies}
\format{character vector}
\usage{
absts_ca1009
}
\description{
vector of character strings representing study.abstracts from the 1009 cancer studies
}
\note{
11 studies had no abstract and the title of the study is substituted.
}
\keyword{datasets}
