% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bag_fields_ca1009}
\alias{bag_fields_ca1009}
\title{names of fields used in 1009 cancer studies}
\format{named vector}
\usage{
bag_fields_ca1009
}
\description{
names of fields used in 1009 cancer studies
}
\keyword{datasets}
