% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdc_dx_sam}
\alias{gdc_dx_sam}
\title{tibble with selected information from Genomic Data Commons dictionary}
\format{tibble}
\usage{
gdc_dx_sam
}
\description{
tibble with selected information from Genomic Data Commons dictionary
}
\keyword{datasets}
