% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontolook.R
\name{load_ontolookup}
\alias{load_ontolookup}
\title{retrieve a snapshot of terms in ORDO HP UBERON DOID NCIT}
\usage{
load_ontolookup(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{instance of `BiocFileCache`, defaults to `BiocFileCache::BiocFileCache()`}
}
\value{
a data.frame
}
\description{
retrieve a snapshot of terms in ORDO HP UBERON DOID NCIT
}
\examples{
if (interactive()) head(load_ontolookup())
}
