% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontolook.R
\name{make_ontolookup}
\alias{make_ontolookup}
\title{create through downloads a tibble amalgamating terms in ORDO HP UBERON DOID NCIT}
\usage{
make_ontolookup()
}
\value{
tibble
}
\description{
create through downloads a tibble amalgamating terms in ORDO HP UBERON DOID NCIT
}
\note{
`load_ontolookup` will retrieve an April 2019 snapshot of the data in a tibble from AWS S3 or local cache
}
\examples{
if (interactive()) {
  if (isTRUE(askYesNo("This will perform a download of 5 ontologies including NCIT, are you sure you want to proceed?"))) {
      cur_ontolook = make_ontolookup()
    cur_ontolook
      }
 }
}
\author{
Sean Davis, minor mods by VJ Carey
}
