% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCa1009.R
\name{putCa1009}
\alias{putCa1009}
\title{retrieve metadata examples from AWS S3 to BiocFileCache}
\usage{
putCa1009(cache = BiocFileCache::BiocFileCache())
}
\description{
retrieve metadata examples from AWS S3 to BiocFileCache
}
