% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{study_publ_dates}
\alias{study_publ_dates}
\title{NCBI SRA publication date information for 6212 studies}
\format{data.frame}
\usage{
study_publ_dates
}
\description{
NCBI SRA publication date information for 6212 studies
}
\keyword{datasets}
